requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
djangorestframework==3.16.0
frozenlist==1.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
twilio==9.6.5
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
zope.interface==7.2
